+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Bugfixes
 
 * Fix `#connection_completed` not be called after `#connect`.
+* Raise if another `#reconnect` is under progress.
 
 # v0.1.0